pkg/endpoint: move SetStateLocked to state.go
Signed-off by: Ian Vernon <ian@cilium.io>
